Gui Updating
1. Divide Wrapper class to several one
2. Rework Rendering
