+++ editor/incl/editor/Gui/Util.hpp
@@ -1,0 +1,7 @@
+#pragma once
+
+#include <imgui/imgui.h>
+
+namespace ImGui {
+    ImTextureID convertTextureHandle(const unsigned int& handle);
+}
