+++ editor/incl/editor/ScenePanel.hpp
@@ -35,7 +35,7 @@ namespace editor {
         ScenePanel();
         ~ScenePanel()override = default;
 
-        void setActiveScene(Scene::Ptr&& ptr) { m_scene = ptr; m_selectedEntity = {};}
+        void setActiveScene(Scene::Ptr ptr) { m_scene = ptr; m_selectedEntity = {};}
         void render() override;
     private:
         void drawEntity(robot2D::ecs::Entity entity);
